+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+0.2.0.0: [2018.07.22]
+---------------------
+* Fixed the `Semiring` instances of `Set`, `HashSet`, `Vector`, `Storable Vector`, `Unboxed Vector`.
+* Removed the `Semiring` instances of `Seq`, `Alt`, `Endo`.
+* Added comprehensive test suite that tests all `Semiring` instances defined in Data.Semiring
+* Added Free semiring
+* Added newtypes: `Add`, `Mul`
+
 0.1.2: [2018.05.04]
 -------------------
 * `semirings` now builds back to